[blur][ncl] Update blurView demo screen
